[DIDA] feat. 커뮤니티 상세 > 커뮤니티 상세 -> NFT상세 이동 로직 추가

+++ domain/src/main/java/com/dida/domain/model/nav/post/Post.kt
@@ -2,7 +2,7 @@ package com.dida.domain.model.nav.post
 
 data class Post(
     val postId: Long,
-    val cardId: Int,
+    val cardId: Long,
     val userId: Long,
     val userName: String,
     val userImgUrl: String,
